refactor: Moved leaderboard to \gameSession
